+++ src/main/java/com/promesa/promesa/domain/review/dto/response/ReviewResponse.java
@@ -1,6 +1,5 @@
 package com.promesa.promesa.domain.review.dto.response;
 
-import com.promesa.promesa.common.application.S3Service;
 import com.promesa.promesa.domain.review.domain.Review;
 
 import java.util.List;
@@ -13,16 +12,14 @@ public record ReviewResponse(
         int rating,
         List<String> reviewImages
 ){
-    public static ReviewResponse from(Review review, S3Service s3Service, String bucket) {
+    public static ReviewResponse from(Review review, List<String> imageKeys) {
         return new ReviewResponse(
                 review.getId(),
                 review.getContent(),
                 review.getItem().getId(),
                 review.getMember().getId(),
                 review.getRating(),
-                review.getReviewImages().stream()
-                        .map(image -> s3Service.createPresignedGetUrl(bucket, image.getKey()))
-                        .toList()
+                imageKeys
         );
     }
 }
